Move ResultCount to a separate component
