+++ extensions/ql-vscode/src/view/results/ResultCount.tsx
@@ -1,0 +1,29 @@
+import * as React from "react";
+import { ResultSet } from "../../common/interface-types";
+import { tableHeaderItemClassName } from "./result-table-utils";
+
+interface Props {
+  resultSet?: ResultSet;
+}
+
+function getResultCount(resultSet: ResultSet): number {
+  switch (resultSet.t) {
+    case "RawResultSet":
+      return resultSet.schema.rows;
+    case "InterpretedResultSet":
+      return resultSet.interpretation.numTotalResults;
+  }
+}
+
+export function ResultCount(props: Props): JSX.Element | null {
+  if (!props.resultSet) {
+    return null;
+  }
+
+  const resultCount = getResultCount(props.resultSet);
+  return (
+    <span className={tableHeaderItemClassName}>
+      {resultCount} {resultCount === 1 ? "result" : "results"}
+    </span>
+  );
+}
